+++ src/types/Application.d.ts
@@ -158,12 +158,3 @@ type Organization = {
   _id: string;
   name: string;
 };
-
-type OrgInfo = {
-  orgID: string;
-  orgName: string;
-  orgRole: "Owner" | "Submitter" | "Concierge"; // Concierge can only be assign to a Curator
-  orgStatus: "Active" | "Inactive" | "Disabled";
-  createdAt: string; // 2023-05-01T09:23:30Z, ISO data time format
-  updateAt: string; // 2023-05-01T09:23:30Z  ISO data time format
-};
