+++ src/types/Auth.d.ts
@@ -15,3 +15,12 @@ type UserInput = {
   firstName: string;
   lastName: string;
 };
+
+type OrgInfo = {
+  orgID: string;
+  orgName: string;
+  orgRole: "Owner" | "Submitter" | "Concierge"; // Concierge can only be assigned to a Curator
+  orgStatus: "Active" | "Inactive" | "Disabled";
+  createdAt: string; // 2023-05-01T09:23:30Z, ISO data time format
+  updateAt: string; // 2023-05-01T09:23:30Z  ISO data time format
+};
